Task_6

+++ src/main/java/pl/futurecollars/invoicing/service/InvoiceService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.UUID;
 import org.springframework.stereotype.Service;
 import pl.futurecollars.invoicing.db.Database;
+import pl.futurecollars.invoicing.exceptions.InvoiceNotFoundException;
 import pl.futurecollars.invoicing.model.Invoice;
 
 @Service
@@ -31,8 +32,12 @@ public class InvoiceService {
         return database.update(invoice);
     }
 
-    public void deleteInvoice(UUID id) {
-        database.delete(id);
+    public void deleteInvoice(UUID id) throws InvoiceNotFoundException {
+        if (database.checkIfInvoiceExist(id)) {
+            database.delete(id);
+        } else {
+            throw new InvoiceNotFoundException("Invoice with id %s does not exists in db" + id);
+        }
     }
 
 }
